# Replace the internal sha512 implementation with .Net built in sha512