… src/NATS.Client.Core/NaCl/Internal/Array16.cs
… src/NATS.Client.Core/NaCl/Internal/Array8.cs
… src/NATS.Client.Core/NaCl/Internal/ByteIntegerConverter.cs
… src/NATS.Client.Core/NaCl/Internal/Sha512Internal.cs