This pr replaces the internal implementation of Sha512 to use the Sha512 in the .Net framwork.

It also removes the files that were there only to support the internal Sha512 implementation.

The benefits are:

* Simplification by removing code not needed to be maintained
* Easier path to e.g. remove allocations in Nkey because of the methods that supports spans etc.
* Less allocations
* Speed

## Removed methods

This pr also removes some methods from the `Sha512` internal wrapper class, they were all unused internal methods.

```csharp
public void Update(ArraySegment<byte> data)
public void Finalize(ArraySegment<byte> output)
public static byte[] Hash(byte[] data)
```

## Micro benchmarks

```text
BenchmarkDotNet v0.13.12, macOS Sonoma 14.3.1 (23D60) [Darwin 23.3.0]
Apple M2 Pro, 1 CPU, 12 logical and 12 physical cores
.NET SDK 8.0.101
  [Host]   : .NET 8.0.1 (8.0.123.58001), Arm64 RyuJIT AdvSIMD
  ShortRun : .NET 8.0.1 (8.0.123.58001), Arm64 RyuJIT AdvSIMD

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

Main
| Method     | Iter | Mean     | Error   | StdDev  | Gen0     | Allocated |
|----------- |----- |---------:|--------:|--------:|---------:|----------:|
| NKeyCreate | 5000 | 142.9 ms | 5.23 ms | 0.29 ms | 500.0000 |   4.08 MB |

This pr
| Method     | Iter | Mean     | Error   | StdDev  | Gen0     | Allocated |
|----------- |----- |---------:|--------:|--------:|---------:|----------:|
| NKeyCreate | 5000 | 142.8 ms | 2.23 ms | 0.12 ms | 250.0000 |    2.9 MB |
 
Main 
| Method     | Iter | DataSize | Mean      | Error    | StdDev   | Gen0     | Allocated |
|----------- |----- |--------- |----------:|---------:|---------:|---------:|----------:|
| Sha512Hash | 5000 | 1024     |  17.95 ms | 0.346 ms | 0.019 ms | 187.5000 |    1.6 MB |
| Sha512Hash | 5000 | 25600    | 396.04 ms | 4.489 ms | 0.246 ms |        - |    1.6 MB |

This pr
| Method     | Iter | DataSize | Mean      | Error     | StdDev    | Gen0    | Allocated |
|----------- |----- |--------- |----------:|----------:|----------:|--------:|----------:|
| Sha512Hash | 5000 | 1024     |  4.991 ms | 0.3914 ms | 0.0215 ms | 46.8750 | 429.69 KB |
| Sha512Hash | 5000 | 25600    | 84.237 ms | 1.3626 ms | 0.0747 ms |       - | 429.81 KB |

```