@mtmk 2024-02-10
> @mtmk is this a flapper: https://github.com/nats-io/nats.net.v2/actions/runs/7855013069/job/21436316896?pr=390#step:7:302
> 
> It's the `Slow_consumer` test, and seems to work fine locally.

yes. we have #344 to take care of them soon.